+++ braindecode/models/__init__.py
@@ -10,5 +10,6 @@ from .eegresnet import EEGResNet
 from .tcn import TCN
 from .sleep_stager_chambon_2018 import SleepStagerChambon2018
 from .tidnet import TIDNet
+from .usleep import USleep
 from .util import get_output_shape
 from .modules import TimeDistributed