+++ docs/whats_new.rst
@@ -70,6 +70,7 @@ Enhancements
 - Removing code duplicate in BaseDataset and WindowsDataset (:gh:`159` by `Lukas Gemein`_)
 - Only load data if needed during preprocessing (e.g., allow timecrop without loading) (:gh:`164` by `Robin Tibor Schirrmeister`_)
 - Adding option to sort filtered channels by frequency band for the filterbank in :func:`braindecode.datautil.filterbank` (:gh:`185` by `Lukas Gemein`_)
+- Adding the USleep model :class:`braindecode.models.USleep` (:gh:`282` by `Theo Gnassounou`_ and `Omar Chehab`_)
 
 Bugs
 ~~~~
@@ -100,6 +101,8 @@ Authors
 .. _Dan Wilson: https://github.com/dcwil
 .. _Henrik Bonsmann: https://github.com/HenrikBons
 .. _Vytautas Jankauskas: https://github.com/vytjan
+.. _Theo Gnassounou: https://github.com/Tgnassou
+.. _Omar Chehab: https://github.com/l-omar-chehab
 .. _Alex Gramfort: http://alexandre.gramfort.net
 .. _Cedric Rommel: https://cedricrommel.github.io
 .. _Simon Brandt: https://github.com/sbbrandt